Chat can now redeem points to provide their custom color hex code
Chat can now use !light <username> || me after they redeem their points and send in a valid hex code to change the color to that hex code.
Reading/Writing to a file is now more efficient with the new method that @thatN00b__ has provided.
Added ability to set the lights to the stored hex code if the command !light <username> || me was used.